requirements.txt
streamlit
pandas==2.1.4
python-dotenv
requests